requirements.txt
# Flask Core
Flask==3.0.0
Werkzeug==3.0.1

# Flask Extensions
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5

# Database
SQLAlchemy==2.0.23
PyMySQL==1.1.0

# Forms & Validation
WTForms==3.1.1
email-validator==2.3.0

# Image Processing
Pillow==10.1.0

# Environment & Configuration
python-dotenv==1.0.0

# HTTP Requests (for currency API)
requests==2.31.0

# Security & Utilities
cryptography==41.0.7

# CSV Import
chardet==5.2.0
Flask-Executor==1.0.0

# Development Tools (optional)
python-dateutil==2.8.2

# WSGI Server for Production
gunicorn==21.2.0